% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{analyze_binary}
\alias{analyze_binary}
\title{Analyze binary fit.}
\usage{
analyze_binary(data, classes, ci = FALSE)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{classes}{a vector with classes of the outcome.}
}
\value{
a tibble with the statistic values.
}
\description{
computes fit errors, C-index, kappa and ROC for binary classification.
}
