% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{analyze_cv}
\alias{analyze_cv}
\title{Get CV prediction statistics.}
\usage{
analyze_cv(data, fun, ci_method = c("percentile", "bca", "norm"), ...)
}
\arguments{
\item{data}{a data frame with .outcome, .fitted and .resample columns.}

\item{fun}{analysis function.}

\item{ci_method}{method used for 95\% CI calculation.}

\item{...}{additional arguments passed to the analysis function.}
}
\value{
a tibble with the statistic values.
}
\description{
Computes fit statistics for each CV fold and returns the statistics' means
with 95\% confidence intervals (CI).
}
