% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{analyze_multi}
\alias{analyze_multi}
\title{Analyze multi-class fit.}
\usage{
analyze_multi(data, ci = FALSE)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{classes}{a vector with classes of the outcome.}
}
\value{
a tibble with the statistic values.
}
\description{
computes accuracy and kappa.
}
