% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{analyze_reg}
\alias{analyze_reg}
\title{Analyze regression fit.}
\usage{
analyze_reg(data, ci = FALSE)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{ci}{should 95\% CI be computed?}
}
\value{
a tibble with the statistic values.
}
\description{
computes fit errors and fit:outcome correlations.
}
