% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{augment.caretx}
\alias{augment.caretx}
\title{Predictions with explanatory variables.}
\usage{
\method{augment}{caretx}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{caretx}} object.}

\item{newdata}{an optional data frame with the test data.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of data frames. Each of them contains the
observation number (\code{.observation}),
resample ID (only for resample/CV, \code{.resample}),
outcome (\code{.outcome}), fitted values/classes (\code{.fitted}).
For classification models, class assignment probabilities are returned
as well (columns named after levels of the \code{.outcome} variable).
}
\description{
The \code{augment()} method for \code{\link{caretx}} objects derives the
predictions for the training, resample (CV), and, optionally, test
data set appended with the explanatory variables.
}
