% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration.predx}
\alias{calibration.predx}
\alias{calibration.caretx}
\title{Calibrate predictions and models.}
\usage{
\method{calibration}{predx}(
  x,
  bs = "cr",
  k = 20,
  qu = 0.5,
  form = NULL,
  lsig = NULL,
  err = NULL,
  control = list(link = "identity"),
  argGam = NULL,
  ...
)

\method{calibration}{caretx}(
  x,
  newdata = NULL,
  bs = "cr",
  k = 20,
  qu = 0.5,
  form = NULL,
  lsig = NULL,
  err = NULL,
  control = list(link = "identity"),
  argGam = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{\link{caretx}} model or \code{\link{predx}} prediction
object.}

\item{bs}{basis function for the smoother, ignored if a formula provided.}

\item{k}{degrees of freedom for the smoother, ignored if a formula provided.}

\item{qu}{quantile or quantiles for the calibration,
see: \code{\link[qgam]{qgam}} for details.}

\item{form}{optional GAM formula as specified by
\code{\link[mgcv]{formula.gam}}. The uncalibrated predictions are stored
internally in the '.raw' variable, which needs to be included in the
user-provided formula.}

\item{lsig}{the value of the log learning rate used to create the
Gibbs posterior, see: \code{\link[qgam]{qgam}} for details.}

\item{err}{an upper bound on the error of the estimated quantile curve,
see: \code{\link[qgam]{qgam}} for details.}

\item{control}{a list of control parameters passed to
\code{\link[qgam]{qgam}}.}

\item{argGam}{a list of parameters to be passed to
\code{\link[mgcv]{gam}}, with the exception of formula, family and data.}

\item{...}{extra arguments passed to \code{\link[qgam]{qgam}}.}

\item{newdata}{a data frame with the test data.#'}
}
\value{
A list with the \code{predx} object
(.raw stores the uncalibrated predctions, .fitted stores the calibrated
predictions) along with the gamObject named \code{cal_fit},
the chosen quantile value (\code{qu}) and values of explained deviance (\code{qu_tbl}).
}
\description{
Enables post-hoc quantile GAM calibration of the regression predictions
and models using the \code{\link[qgam]{qgam}} tool.
}
\details{
Currently, only quantile calibration for regression \code{predx} and \code{caretx}
objects is implemented. If a vector of quantiles is provided, the optimal
one is choosen based on the minimum RMSE
(more formal criteria are in development).
}
