% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{classp}
\alias{classp}
\alias{classp.predx}
\alias{classp.caretx}
\title{Extract class assignment probability from models and predictions.}
\usage{
classp(x, ...)

\method{classp}{predx}(x, ...)

\method{classp}{caretx}(x, newdata, ...)
}
\arguments{
\item{x}{\code{\link{caretx}} model or \code{\link{predx}} prediction object.}

\item{...}{extra arguments passed to methods, currently none.}

\item{newdata}{test data set (optional).}
}
\value{
For \code{predx} objects, a data frame with
the columns \code{.outcome} and \code{.fitted} representing the true outcome
and predicted class assignment, respectively, columns storing class
assignment probabilities as welll as the \code{winner_p} variable storing
the probability for the predicted class.
For \code{caretx} models, the function returns a list of tibbles described above,
each one for the training, resample (CV) and test data sets.
}
\description{
Computes the class assignment probability for classification models.
}
\details{
Returns NULL for regression models and predictions and throws a warning.
The \code{classp} is a S3 generic function.
}
