% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_stats.R
\name{clplots}
\alias{clplots}
\alias{clplots.predx}
\alias{clplots.caretx}
\title{Receiver operating characteristic plots for outcome classes.}
\usage{
clplots(x, ...)

\method{clplots}{predx}(x, one_plot = TRUE, ...)

\method{clplots}{caretx}(x, newdata = NULL, one_plot = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{predx}} prediction object or a \code{\link{caretx}}
model.}

\item{...}{extra arguments passed to \code{\link{plot_class_roc}}.}

\item{one_plot}{logical: should all ROC curves be displayed in one plot?}

\item{newdata}{optional, a data frame with the test data.}
}
\value{
a single \code{ggplot} object or a list of \code{ggplot} objects.
For \code{clplots.caretex} a list of \code{ggplots} with plots for the training data,
resamples and, if \code{newdata} is specified, also for the test data set.
}
\description{
Draws receiver operating characteristic (ROC) plots for particular outcome
classes. The ROC statistics for particular classes are obtained by comparing
the given class the remaining ones (one versus rest comparison).
}
\details{
The function employs internally \code{\link[caret]{multiClassSummary}} and
plotting tools from the \code{plotROC} package. \code{clplots} is a S3 generic
function.
}
