% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_stats.R
\name{clstats}
\alias{clstats}
\alias{clstats.predx}
\alias{clstats.caretx}
\title{Class detection quality.}
\usage{
clstats(x, ...)

\method{clstats}{predx}(x, ...)

\method{clstats}{caretx}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{predx}} prediction object or a \code{\link{caretx}}
model.}

\item{...}{arguments passed to methods, currently none.}

\item{newdata}{optional, a data frame with the test data.}
}
\value{
For \code{predx} objects: a data frame with class-specific
receiver-operator characteristic as well as class-specific Brier scores
and average class assignment probabilities.
For \code{caretx} a list of such data frames, each one for the train, resample
and train data set.
}
\description{
Computes a bunch of numeric statistics for quality of identification of
particular classes in an one versus rest manner.
}
\details{
For regression, NULL is returned with a warning.
The function uses internally \code{\link[caret]{multiClassSummary}}.
\code{clstats} is a S3 generic function.
}
\references{
Kuhn M. Building predictive models in R using the caret package.
J Stat Softw (2008) 28:1–26. doi:10.18637/jss.v028.i05

Brier GW. VERIFICATION OF FORECASTS EXPRESSED IN TERMS OF PROBABILITY.
Mon Weather Rev (1950) 78:1–3.
doi:10.1175/1520-0493(1950)078<0001:vofeit>2.0.co;2

Goldstein-Greenwood J. A Brief on Brier Scores | UVA Library. (2021) Available
at: https://library.virginia.edu/data/articles/a-brief-on-brier-scores

Cohen J. A Coefficient of Agreement for Nominal Scales. Educ Psychol Meas
(1960) 20:37–46. doi:10.1177/001316446002000104
}
