% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{collapse_cv}
\alias{collapse_cv}
\alias{collapse_cv.predx}
\alias{collapse_cv.caretx}
\title{Case-wise collapsing of out-of-fold predictions.}
\usage{
collapse_cv(x, ...)

\method{collapse_cv}{predx}(x, fun = mean, ...)

\method{collapse_cv}{caretx}(x, fun = mean, newdata = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{predx}} object with model's predictions or
a \code{\link{caretx}} model.}

\item{...}{additional arguments passed to \code{fun}.}

\item{fun}{a function used for aggregation of predicted values (regression)
or class probabilities. It should take a numeric vector and return
a single numeric value. \code{mean()} by default.}

\item{newdata}{test data set (optional).}
}
\value{
The \code{collapse_cv} method for \code{\link{predx}} objects returns an
\code{\link{predx}} with the collapsed observations.
The \code{collapse_cv} method for \code{\link{caretx}} models generates predictions
for the training, CV, and optionally, test data, and returns a list of
\code{\link{predx}} objects.
}
\description{
Function \code{collapse_cv()} collapses out-of-fold predictions in cross-validation
by an user-provided function like \code{mean()} or \code{median()} in an observation-wise
manner.
}
\details{
The function collapses only predictions made in repeated cross-validation,
objects storing predictions of canonical cross-validation, or objects with
predictions in the training or test data are returned without any modifications.
For predictions made by regression, the predicted values are simply aggregated
by function \code{fun}.
For predictions made by classification models, the class-assignment probabilities
are aggregated by function \code{fun}, and the class assignment is determined by
nodal voting (see \code{\link{nodal_voting}}).
Because there is no possibility to elucidate the cross-validation fold identifier
for the aggregated observations, the \code{.resample} column in the output
\code{\link{predx}} object is replaced with the fold-assignment in the
first repeat of the cross-validation.
The function can't handle \code{predx} objects with calibration function attached:
in such cases, no modifications are made, and a warning is raised.
}
