% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{comp_class_stats}
\alias{comp_class_stats}
\alias{comp_class_roc}
\title{Accuracy, assignment probability, Brier scores and ROC for outcome classes.}
\usage{
comp_class_stats(predx_object)

comp_class_roc(predx_object)
}
\arguments{
\item{predx_object}{a \code{predx} class object.}
}
\value{
a data frame with columns representing the requested statistics for
particular classes.
}
\description{
Those internally utilized functions compute:
\itemize{
\item \code{comp_class_stats()}:  average class assignment probability and
Brier scores for the outcome classes.
\item \code{comp_class_roc()}: accuracy, sensitivity, specificity, recall etc.
for each of the class compared with the rest.
}
}
\details{
For regression, NULL is returned with a warning.
\code{comp_class_roc()} uses \code{\link[caret]{multiClassSummary}}.
}
