% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{components.caretx}
\alias{components.caretx}
\alias{components.predx}
\alias{model.frame.predx}
\alias{model.frame.caretx}
\title{Access features and statistics of models and predictions}
\usage{
\method{components}{caretx}(
  object,
  newdata = NULL,
  what = c("data", "formula", "resample", "tuning", "best_tune", "prediction",
    "probability", "square_dist", "fit", "n", "n_classes", "residuals", "confusion"),
  ...
)

\method{components}{predx}(
  object,
  what = c("data", "prediction", "classes", "probability", "square_dist", "fit", "n",
    "n_classes", "residuals", "confusion"),
  ...
)

\method{model.frame}{predx}(formula, ...)

\method{model.frame}{caretx}(formula, ...)
}
\arguments{
\item{object}{a \code{\link{caretx}} model or
\code{\link{predx}} prediction object.}

\item{newdata}{test data set required for making predictions.
NULL by default, used only if \code{what} is set to \code{prediction}.}

\item{what}{name of the requested feature:
\itemize{
\item \code{data}: the training data (\code{caretx}) or predictions (\code{predx})
\item \code{formula}: the model's formula
\item \code{resample}: a data frame with tuning fit stats in the model's
resamples (e.g. CV folds)
\item \code{tuning}: a data frame with tuning results.
\item \code{best_tune}: the combination of tuning parameters for
the model with the optimal performance in the resamples.
\item \code{prediction}: the predictions, optionally for the test data provides as
the \code{newdata} argument
\item \code{fit}: returns statistics of the model fit
\item \code{classes}: returns the outcome class vector
\item \code{probability}: a data frame or a list of data frames with class assignment
probability
\item \code{square_dist}: squared distance from the outcome as defined by Brier at al.
\item \code{n}: the complete case number
\item \code{n_classes}: number of observations in the true outcome and fitted classes.
NULL for regression models.
\item \code{residuals}: model fit residuals
\item \code{confusion}: computes the confusion matrix with observation counts
}}

\item{...}{additional arguments passed to the specific methods.}

\item{formula}{\code{\link{caretx}} model or
\code{\link{predx}} prediction object.}
}
\value{
the requested feature.
}
\description{
The \code{components()} and \code{model.frame()} methods provide handy extractor
functions enabling access to the data, predictions, number of observations
and fit summary.
}
\references{
Brier GW. VERIFICATION OF FORECASTS EXPRESSED IN TERMS OF PROBABILITY.
Mon Weather Rev (1950) 78:1–3.
doi:10.1175/1520-0493(1950)078<0001:vofeit>2.0.co;2

Goldstein-Greenwood J. A Brief on Brier Scores | UVA Library. (2021) Available
at: https://library.virginia.edu/data/articles/a-brief-on-brier-scores
}
