% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{confusion.caretx}
\alias{confusion.caretx}
\title{Confusion matrix for binary and multi-class caretx models.}
\usage{
\method{confusion}{caretx}(
  caretx_model,
  newdata = NULL,
  scale = c("none", "fraction", "percent")
)
}
\arguments{
\item{caretx_model}{caretx model.}

\item{newdata}{newdata test data set.}

\item{scale}{indicates, how the table is to be scales. 'none' returns the counts, 'fraction' returns the fraction
of all observations, 'percent' returns the percent of all observations. Defaults to 'none'.}
}
\value{
a list of table object storing the confusion matrices. For regression models NULL and a warning is generated.
}
\description{
Creates a confusion matrix, valid only for binary and multi-class predictions.
}
