% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_oop.R
\name{confusion.predx}
\alias{confusion.predx}
\title{Confusion matrix for binary and multi-class predx objects.}
\usage{
\method{confusion}{predx}(predx_object, scale = c("none", "fraction", "percent"))
}
\arguments{
\item{predx_object}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}

\item{scale}{indicates, how the table is to be scales. 'none' returns the counts, 'fraction' returns the fraction
of all observations, 'percent' returns the percent of all observations. Defaults to 'none'.}
}
\value{
a table object storing the confusion matrix. For regression models NULL and a warning is generated.
}
\description{
Creates a confusion matrix, valid only for binary and multi-class predictions.
}
\details{
a wrapper around the base \code{\link[base]{table}} function.
}
