% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{confusion}
\alias{confusion}
\alias{confusion.caretx}
\alias{confusion.predx}
\title{Confusion matrix for binary and multi-class models and predictions.}
\usage{
confusion(x, ...)

\method{confusion}{caretx}(x, newdata = NULL, scale = c("none", "fraction", "percent"), ...)

\method{confusion}{predx}(x, scale = c("none", "fraction", "percent"), ...)
}
\arguments{
\item{x}{\code{\link{caretx}} model or \code{\link{predx}} prediction object.}

\item{...}{extra arguments passed to methods.}

\item{newdata}{newdata test data set.}

\item{scale}{indicates, how the table is to be scaled:
\itemize{
\item \code{none}: returns the counts (default)
\item \code{fraction} returns the fraction of all observations
\item \code{percent} returns the percent of all observations
}}
}
\value{
a list of table object storing the confusion matrices (\code{caretx})
or a single table bject representing the confusion matrix  (\code{predx}).
For regression models NULL and a warning is generated.#
}
\description{
Creates a confusion matrix, valid only for binary and
multi-class models and predictions.
}
\details{
\code{confusion()} is a S3 generic function.
}
