% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{dummy_}
\alias{dummy_}
\title{Resolves a factor to a dummy matrix.}
\usage{
dummy_(x)
}
\arguments{
\item{x}{a factor.}
}
\value{
a numeric binary (0/1) matrix with dummy coding of the vector's levels.
}
\description{
Resolves a factor to a dummy matrix.
The code is inspired by \code{\link[DescTools]{Dummy}}.
}
