% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{extract.caretx}
\alias{extract.caretx}
\title{Extract features of caretx models.}
\usage{
\method{extract}{caretx}(
  caretx_model,
  newdata = NULL,
  what = c("data", "prediction", "fit", "n", "residuals", "confusion"),
  ...
)
}
\arguments{
\item{caretx_model}{caretx model.}

\item{newdata}{newdata test data set.}

\item{what}{name of the requested feature.}

\item{...}{additional arguments passed to the specific methods.}
}
\value{
the requested feature.
}
\description{
a handy extractor function enabling access to the data, predictions, number of observations and fit summary.
}
\details{
'data' returns the training data set, 'prediction' returns the predicted values,
'fit' the summary of fit stats, 'n' the number of complete cases, 'residuals' returns the fit residuals,
'confusion' extracts the confusion matrices with counts.
}
