% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_brier}
\alias{get_brier}
\alias{get_classp}
\title{Calculate Brier score and weighted class probability average.}
\usage{
get_brier(data, ci_method = c("percentile", "bca", "norm"))

get_classp(data, ci_method = c("percentile", "bca", "norm"))
}
\arguments{
\item{data}{a data frame with the columns \code{.outcome}, \code{.fitted},
and, optionally \code{.resample}. The column \code{square_dist} is required for the
Brier scores. The column \code{winner_p} is required for the average probabilites.}

\item{ci_method}{method used for 95\% CI calculation.}
}
\value{
a tibble with the columns \code{statistic}, \code{estimate}, \code{lower_ci} and
\code{upper_ci} storing the statistic name, its estimated value
as well as lower and upper limit of the 95\% confidence interval.
}
\description{
\code{get_brier()} computes the Brier score (statistic name \code{brier_score}).
\code{get_classp()} computes a weighted arithmetic mean of the winning class
assignment probabilities. The weights are frequencies of the true outcome
classes(statistic name \code{class_p_outcome}) and predicted classes
(\code{class_p_fitted}).
For cross-validation, mean metrics with 95\% confidence intervals
are returned.
}
