% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_ci}
\alias{get_ci}
\title{Calculate mean with 95\% confidence interval}
\usage{
get_ci(vector, method = c("percentile", "bca", "norm"))
}
\arguments{
\item{vector}{a numeric vector.}

\item{method}{the method used for calculation of 95\% confidence interval.}
}
\value{
a data frame with mean, lower CI and upper CI values
}
\description{
Computes mean with 95\% confidence interval.#'
}
