% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_class_error}
\alias{get_class_error}
\title{Calculates classification error.}
\usage{
get_class_error(data)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}
}
\value{
a tibble with the statistic values.
}
\description{
Calculates simple percentage of correct prediction and failures.
}
