% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_correlations}
\alias{get_correlations}
\title{Calculate correlations between the fitted and outcome values.}
\usage{
get_correlations(data, ci = TRUE)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{ci}{logical, should 95\% confidence intervals be calculated?}
}
\value{
a tibble with the statistic values.
}
\description{
calculates Pearson, Spearman and Kendal TauB correlations between the fitted and outcome values.
}
