% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_errors}
\alias{get_errors}
\title{Calculate fit errors.}
\usage{
get_errors(data)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}
}
\value{
a tibble with the statistic values.
}
\description{
calculates MAE, MSE, RMSE and pseudo R squared based on MSE and outcome variance.
}
