% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qc.R
\name{get_qc_plots}
\alias{get_qc_plots}
\title{Diagnostic plots of model residuals for predx objects.}
\usage{
get_qc_plots(
  predx_object,
  cust_theme = ggplot2::theme_classic(),
  point_wjitter = 0.01,
  point_hjitter = 0.01,
  point_alpha = 0.75
)
}
\arguments{
\item{predx_object}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{point_wjitter}{horizontal jittering of the points.}

\item{point_hjitter}{vertical jittering of the points.}

\item{point_alpha}{plot point alpha.}
}
\value{
returns a series of ggplot objects with the diagnostic residuals plots.
}
\description{
Plots a series of diagnostic plots of model residuals with potential outliers indicated.
}
