% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qc.R
\name{get_qc_tbl}
\alias{get_qc_tbl}
\title{Extended residuals of a predx object.}
\usage{
get_qc_tbl(predx_object)
}
\arguments{
\item{predx_object}{an object of class \code{predx}.}
}
\value{
a data frame with the fitted values, true outcome,
residuals and candidate outliers.
}
\description{
Extracts extended working residuals of prediction and potential outliers.
}
\details{
Calculates working residuals for regression and binary
classification predictions.
In addition, squared and standardized residuals are returned along with
expected normal distribution values for the standardized residuals
and the true outcome.
Potential outliers are identified by the two-SD criterion.
}
