% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_resids}
\alias{get_resids}
\title{Calculate residuals.}
\usage{
get_resids(data)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}
}
\value{
a numeric vector of residuals.
}
\description{
calculates working residuals.
}
