% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_resids_binary}
\alias{get_resids_binary}
\alias{get_var_binary}
\alias{get_errors_binary}
\title{Binary classifier residuals, variance and errors.}
\usage{
get_resids_binary(data, classes)

get_var_binary(data)

get_errors_binary(data, classes)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{classes}{a vector with the class order.}
}
\value{
a numeric or a  tibble with the statistic values.
}
\description{
Retrieves working binary residuals (\code{get_resids_binary()}),
outcome variance (\code{get_var_binary()}) and fit errors
(\code{get_errors_binary()}: MAE, MSE, RMSE and pseudo-R-squared).
}
