% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_roc}
\alias{get_roc}
\title{Calculates simple receiver-operating characteristic.}
\usage{
get_roc(data, classes)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{classes}{a vector with the class order.}
}
\value{
a tibble with the statistic values.
}
\description{
A wrapper around \code{\link[caret]{twoClassSummary}}.
}
