% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_var}
\alias{get_var}
\title{Calculate outcome variance.}
\usage{
get_var(data)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}
}
\value{
numeric, outcome variance.
}
\description{
calculates outcome variance.
}
