% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_var_binary}
\alias{get_var_binary}
\title{Binary classifier outcome variance.}
\usage{
get_var_binary(data)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}
}
\value{
numeric, outcome variance.
}
\description{
calculates outcome variance.
}
