% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{is_caretx}
\alias{is_caretx}
\title{Check the caretx class.}
\usage{
is_caretx(x)
}
\arguments{
\item{x}{an object.}
}
\value{
A logical value.
}
\description{
Checks if the model is an instance of the caretx class.
}
