% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{is_predx}
\alias{is_predx}
\title{Check the predx class.}
\usage{
is_predx(x)
}
\arguments{
\item{x}{an object.}
}
\value{
A logical value.
}
\description{
Checks if the model is an instance of the predx class.
}
