% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{ngroups.predx}
\alias{ngroups.predx}
\alias{ngroups.caretx}
\title{Numbers of observation in the actual and predicted classes.}
\usage{
\method{ngroups}{predx}(x, ...)

\method{ngroups}{caretx}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{caretx}} model or \code{\link{predx}}
prediction object.}

\item{...}{extra arguments, currently none.}

\item{newdata}{optional test data set.}
}
\value{
a data frame with the class names and numbers of observations
in the outcome and fitted classes (\code{predx}) or a list of such data frames
for the training, resample (CV) and test datasets.
}
\description{
Computes the number of observation in the outcome an predicted classes.
}
\details{
For regression, NULL is returned with a warning.
}
