% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{nobs.caretx}
\alias{nobs.caretx}
\alias{nobs.predx}
\title{Number of complete observations in a model or prediction.}
\usage{
\method{nobs}{caretx}(object, ...)

\method{nobs}{predx}(object, ...)
}
\arguments{
\item{object}{a \code{\link{caretx}} model or \code{\link{predx}}
prediction object.}

\item{...}{extra arguments, currently none.}
}
\value{
number of complete cases used for modeling.
}
\description{
The number of complete observations.
}
