% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_oop.R
\name{nobs.predx}
\alias{nobs.predx}
\title{Number of complete observations in a predx object}
\usage{
\method{nobs}{predx}(predx_object)
}
\arguments{
\item{predx_object}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}
}
\value{
number of complete cases used for modeling.
}
\description{
The number of complete observations.
}
