% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aggregation.R
\name{nodal_voting}
\alias{nodal_voting}
\alias{nodal_voting.default}
\alias{nodal_voting.matrix}
\title{Nodal voting: outcome with the highest probability.}
\usage{
nodal_voting(x, ...)

\method{nodal_voting}{default}(x, ...)

\method{nodal_voting}{matrix}(x, ...)
}
\arguments{
\item{x}{a numeric vector or a matrix.}

\item{...}{arguments passed to the methods, none at the moment.}
}
\value{
an integer value: the index number.
}
\description{
Nodal voting: the function returns the index of a numeric vector, which
corresponds to its maximum.
Ties are resolved at random with a warning.
For matrices, the voting is performed in a row-wise manner.
}
