% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{plot.caretx}
\alias{plot.caretx}
\title{Plot diagnostic plots and predictions versus outcome for caretx models.}
\usage{
\method{plot}{caretx}(
  caretx_model,
  newdata = NULL,
  type = c("diagnostic", "fit", "regression", "roc", "confusion"),
  plot_title = NULL,
  signif_digits = 2,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{caretx_model}{a caretxmodel.}

\item{newdata}{test data set.}

\item{type}{type of the plot. 'diagnostic': diagnostic plots of residuals, 'fit': the fitted versus outcome plot
specific for the prediction character (regression, ROC for binary classification and heat map of the confusion matrix
for multi-class prediction), 'regression': point plot of the fitted vs outcome values, 'roc': ROC plot, 'confusion':
heat map representation of the confusion matrix.}

\item{plot_title}{a vector with plot titles. Length 2 if no newdata provided, otherwise lenght 3.}

\item{signif_digits}{significant digits used for rounding of statistic values presented in the plot.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{extra arguments passed to plotting functions, as specified by the plot type.
For 'diagnostic': \code{\link{get_qc_plots}}, for 'regression': \code{\link{plot_regression}},
for 'roc': \code{\link{plot_roc}}, for 'confusion': \code{\link{plot_confusion}}.}
}
\value{
a list of ggplot objects with the requested content.
}
\description{
Generates diagnostic plots of residuals or plots of fitted versis outcome
as appropriate for the particular prediction type and displays prediction type-specific analysis results.
}
