% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.predx}
\alias{plot.predx}
\alias{plot.caretx}
\title{Draw diagnostic plots and plots for predictions.}
\usage{
\method{plot}{predx}(
  x,
  type = c("diagnostic", "fit", "regression", "roc", "confusion", "class_p",
    "class_stats"),
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  signif_digits = 2,
  cust_theme = ggplot2::theme_classic(),
  ...
)

\method{plot}{caretx}(
  x,
  newdata = NULL,
  type = c("diagnostic", "fit", "regression", "roc", "confusion", "performance",
    "class_p", "class_stats"),
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  signif_digits = 2,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{x}{a \code{\link{caretx}} model or \code{\link{predx}} prediction
object.}

\item{type}{type of the plot:
\itemize{
\item \code{diagnostic}: diagnostic plots of residuals
\item \code{fit}: the fitted versus outcome plot specific for the prediction character
(scatter plot for regression, ROC for binary classification and
heat map of the confusion matrix for multi-class prediction)
\item \code{regression}: scatter plot of the fitted vs outcome values
\item \code{roc}: receiver-operating characteristic (ROC) plot
\item \code{confusion}: heat map representation of the confusion matrix.
\item \code{performance}: a scatter plot with model performance stats in the
test, resample (CV) and, optionally, training data set.
See: \code{\link{plot_performance}} for details.
\item \code{class_p}: a list of scatter plots with square distances to the outcome
(Brier et al.) and class-assignment probabilities. By default, misclassified
observations are labeled. See: \code{\link{plot_class_p}} for details.
\item \code{class_stats}: a list of box plots with square distances to the outcome
(Brier et al.) and class-assignment probabilities.
See: \code{\link{plot_class_stats}} for details.
}}

\item{plot_title}{\code{predx}: plot title;
\code{caretx}: a vector with plot titles, length 2 if no \code{newdata} provided,
otherwise length 3.}

\item{plot_subtitle}{plot subtitle. If not specified by the user,
the sub-title displays statistics specific for the prediction type.}

\item{plot_tag}{plot tag, number of complete observations and events
if not specified otherwise by the user.}

\item{signif_digits}{significant digits used for rounding of statistic
values presented in the plot.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{extra arguments passed to plotting functions, as specified by
the plot type:
\itemize{
\item \code{diagnostic}: \code{\link{get_qc_plots}}
\item \code{regression}: \code{\link{plot_regression}}
\item \code{roc}: \code{\link{plot_roc}}
\item \code{confusion}: \code{\link{plot_confusion}}
\item \code{performance}: \code{\link{plot_performance}}
\item \code{class_p}: \code{\link{plot_class_p}}
\item \code{class_stats}: \code{\link{plot_class_stats}}
}}

\item{newdata}{an optional test data set.}
}
\value{
a ggplot graphic with the requested content (\code{predx}) or a list of
ggplot graphics (\code{caretx}, except for \code{performance} plot type).
}
\description{
Generates diagnostic plots of residuals or plots of fitted versus outcome
as appropriate for the particular prediction type and
displays prediction type-specific analysis results.
}
\references{
Kuhn M. Building predictive models in R using the caret package.
J Stat Softw (2008) 28:1–26. doi:10.18637/jss.v028.i05

Brier GW. VERIFICATION OF FORECASTS EXPRESSED IN TERMS OF PROBABILITY.
Mon Weather Rev (1950) 78:1–3.
doi:10.1175/1520-0493(1950)078<0001:vofeit>2.0.co;2

Goldstein-Greenwood J. A Brief on Brier Scores | UVA Library. (2021) Available
at: https://library.virginia.edu/data/articles/a-brief-on-brier-scores

Cohen J. A Coefficient of Agreement for Nominal Scales. Educ Psychol Meas
(1960) 20:37–46. doi:10.1177/001316446002000104
}
