% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{plot_class_roc}
\alias{plot_class_roc}
\title{Receiver operating characteristic plots for outcome classes.}
\usage{
plot_class_roc(
  predx_object,
  one_plot = FALSE,
  palette = NULL,
  line_size = 0.5,
  point_size = 0.3,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  show_annotation = TRUE,
  annotation_size = 2.75,
  annotation_x = 0.6,
  annotation_y = 0.3,
  annotation_offset = 0.1,
  annotation_hjust = 0,
  annotation_vjust = 1,
  cust_theme = NULL,
  ...
)
}
\arguments{
\item{predx_object}{a \code{predx} class object.}

\item{one_plot}{logical: should all ROC curves be presented in one plot?}

\item{palette}{a character vector with color names of hex codes. If \code{NULL},
the default \code{ggplot2} palette is used.}

\item{line_size}{width of the ROC curve lines.}

\item{point_size}{size of the plot point.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{show_annotation}{logical: should sensitivity and specificity values
be displayed in the plot?}

\item{annotation_size}{size of the annotation text and of the cutoff label.}

\item{annotation_x}{annotation x position.}

\item{annotation_y}{annotation y position.}

\item{annotation_offset}{line spacing in the annotation text.
Applies only if \code{one_plot = TRUE}.}

\item{annotation_hjust}{horizontal justification of the annotation text.}

\item{annotation_vjust}{horizontal justification of the annotation text.}

\item{cust_theme}{custom \code{ggplot} theme. If \code{NULL}, the default ROC theme
of the \code{plotROC} is used.}

\item{...}{extra arguments passed to \code{\link[plotROC]{geom_roc}}.}
}
\value{
if \code{one_plot = TRUE} a single \code{ggplot} object. Otherwise,
a list of \code{ggplot} objects.
}
\description{
Plots receiver-operating characteristic (ROC) plots for outcome classes.
ROC metrics for particular classes are obtained in an one versus rest
comparison.
}
\details{
For regression, NULL is returned with a warning.
}
