% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_confusion}
\alias{plot_confusion}
\title{Plot confusion matrix for a classification predx object.}
\usage{
plot_confusion(
  predx_object,
  scale = c("none", "fraction", "percent"),
  show_labels = TRUE,
  label_size = 2.75,
  label_color = "black",
  signif_digits = 2,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = ".outcome",
  y_lab = ".fitted",
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{predx_object}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}

\item{scale}{indicates, how the table is to be scales. 'none' returns the counts, 'fraction' returns the fraction
of all observations, 'percent' returns the percent of all observations. Defaults to 'none'.}

\item{show_labels}{logical, indicates if counts/fractions/percents are shown in the plot.}

\item{label_size}{size of the label text.}

\item{label_color}{color of the label text.}

\item{signif_digits}{significant digits for rounding of the label values.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag, contains a reference to scale and the observation number if not specified.}

\item{cust_theme}{customized plot theme provided by the user.}
}
\value{
returns a ggplot object.
}
\description{
Generates a heat map representation of the confusion matrix.
Outcome is presented in the X axis, fitted is presented in the Y axis.
}
