% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_performance}
\alias{plot_performance}
\title{Plots of performance metrics in the training, resample and test data.}
\usage{
plot_performance(
  caretx_object,
  newdata = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  show_txt = FALSE,
  txt_size = 2.75,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{caretx_object}{a \code{caretx} object.}

\item{newdata}{optional, a data frame with predictions.}

\item{plot_subtitle}{plot subtitle, see: \code{Details}.}

\item{plot_tag}{plot tag.}

\item{show_txt}{logical, should the data set names (training, CV and test)
be displayed in the plot?}

\item{txt_size}{size of the text labels of the data sets.}

\item{cust_theme}{a custom ggplot theme}

\item{...}{extra arguments passed to \code{\link[ggrepel]{geom_text_repel}}}
}
\value{
a ggplot graphic.
}
\description{
This internally utilized function takes a \code{caretx} model and plots selected
performance stats appropriate for the prediction type in the test, resample
and, optionally, training data set.
Scatter plots are generated, the data set type is color-coded and,
optionally, indicated in the plot.
}
\details{
The plotted performance stats are:
\itemize{
\item for regression: pseudo-R-squared (Y axis) and
root mean square error (RMSE, X axis), Spearman's coefficient for
correlation of the outcome and prediction is represented by the point size.
\item for classification: Brier score (Y axis) and Cohen's kappa, the overall
accuracy is represented by the point size.
}

Numbers of complete observations are displayed in the plot subtitle,
if no user-provided subtitle is specified.
}
