% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_regression}
\alias{plot_regression}
\title{Plot fitted values versus outcome for regression predx objects.}
\usage{
plot_regression(
  predx_object,
  x_var = ".outcome",
  y_var = ".fitted",
  point_size = 2,
  point_shape = 21,
  point_color = "steelblue",
  point_wjitter = 0.01,
  point_hjitter = 0.01,
  point_alpha = 0.75,
  show_trend = TRUE,
  trend_method = "lm",
  show_calibration = TRUE,
  line_size = 0.5,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = x_var,
  y_lab = y_var,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{predx_object}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}

\item{x_var}{name of the variable presented in the X axis.}

\item{y_var}{name of the variable presented in the Y axis.}

\item{point_size}{size of the plot points.}

\item{point_shape}{shape of the points.}

\item{point_wjitter}{horizontal jittering of the points.}

\item{point_hjitter}{vertical jittering of the points.}

\item{point_alpha}{plot point alpha.}

\item{show_trend}{logical, should a trend line be displayed?}

\item{trend_method}{a method for fitting the trend line, see \code{\link[ggplot2]{geom_smooth}} for details, defaults to 'lm'.}

\item{show_calibration}{logical, should a line with slope 1 and intercept 1 be displayed in the plot?}

\item{line_size}{size of the calibration or trend line.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag, number of complete observations if not specified by the user.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{geom_smooth}}.}
}
\value{
returns a  ggplot object.
}
\description{
Generates a point plot with for the fitted and outcome values for
regression predictions with a fitted trend as an option.
}
\details{
The fitted trend is generated by \code{\link[ggplot2]{geom_smooth}}.
}
