% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot a receiver-operator characteristic curve.}
\usage{
plot_roc(
  predx_object,
  line_color = "steelblue",
  line_size = 0.5,
  cutoffs_at = 0.5,
  point_size = 0.3,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  annotation_txt = NULL,
  annotation_color = line_color,
  annotation_size = 2.75,
  annotation_x = 0.6,
  annotation_y = 0.3,
  annotation_hjust = 0,
  annotation_vjust = 1,
  cust_theme = NULL,
  ...
)
}
\arguments{
\item{predx_object}{a \code{predx} object.}

\item{line_color}{color of the ROC line.}

\item{line_size}{size of the ROC line.}

\item{cutoffs_at}{numeric, between 0 and 1, indicates the
cut points to be presented in the ROC curve.}

\item{point_size}{size of the cutoff point.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag, contains the number of complete observations and
vents if not specified by the user.}

\item{annotation_txt}{annotation text.}

\item{annotation_color}{annotation text color.}

\item{annotation_size}{size of the annotation text and of the cutoff label.}

\item{annotation_x}{annotation x position.}

\item{annotation_y}{annotation y position.}

\item{annotation_hjust}{horizontal justification of the annotation text.}

\item{annotation_vjust}{horizontal justification of the annotation text.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{extra arguments passed to \code{\link[plotROC]{geom_roc}}.}
}
\value{
returns a  ggplot object.
}
\description{
Generates a ROC plot with for the fitted and outcome values.
Optionally, a custom annotation inside the plot may be added.
}
\details{
The plot is generated by \code{\link[plotROC]{geom_roc}}.
}
