% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qc.R
\name{point_plot_}
\alias{point_plot_}
\title{Generate a customized scatter plot.}
\usage{
point_plot_(
  data,
  x_var,
  y_var,
  x_lab = x_var,
  y_lab = y_var,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  point_wjitter = 0.01,
  point_hjitter = 0.01,
  point_alpha = 0.75,
  smooth = TRUE,
  silent = TRUE,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{data}{data frame.}

\item{x_var}{name of the variable to be shown in the x axis.}

\item{y_var}{name of the variable to be shown in the y axis.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{point_wjitter}{horizontal jittering of the points.}

\item{point_hjitter}{vertical jittering of the points.}

\item{point_alpha}{plot point alpha.}

\item{smooth}{logical, should a trend line be displayed.}

\item{silent}{logical, display warnings?}

\item{cust_theme}{a custom ggplot theme.}

\item{...}{extra arguments passed to geom_smooth().}
}
\value{
a ggplot graphic
}
\description{
Draws a simple scatter plot for model diagnostic purposes.
}
\details{
As per design, takes the output
of get_qc_tbl() as data argument,
color-codes model potential outliers.
}
