% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{predict.caretx}
\alias{predict.caretx}
\title{Predict the outcome for a caretx model.}
\usage{
\method{predict}{caretx}(object, newdata = NULL, plain = FALSE, ...)
}
\arguments{
\item{object}{\code{\link{caretx}} model.}

\item{newdata}{test data set.}

\item{plain}{logical, should the results be coerced to a single data frame?}

\item{...}{extra arguments, currently none.}
}
\value{
a list of \code{\link{predx}} objects containing the predictions
for the training, resample (CV) and test data sets along with
model and data set type information.
See: \code{\link{components.predx}}, \code{\link{summary.predx}},
\code{\link{residuals.predx}}, \code{\link{confusion.predx}}
and \code{\link{plot.predx}} for useful methods for \code{predx} objects.
If \code{plain} is set to TRUE a prediction data frame if plain is set to TRUE.
}
\description{
Predicts the outcome in train, cross-validation and, optionally,
test data set.
}
\details{
Extends the output of \code{\link[caret]{predict.train}} method.
}
