% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prediction.R
\name{predict_class_cv}
\alias{predict_class_cv}
\title{Predict cross-validation data outcome for classification models.}
\usage{
predict_class_cv(class_model)
}
\value{
a predx object.
}
\description{
a helper function extracting the predictions for the cross-validation data set from
a classification caretx model.
}
