% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prediction.R
\name{predict_class_test}
\alias{predict_class_test}
\title{Predict test data outcome for classification models.}
\usage{
predict_class_test(class_model, newdata)
}
\value{
a predx object.
}
\description{
a helper function extracting the predictions for the test data set from
a classification caretx model.
}
