% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prediction.R
\name{predict_class_train}
\alias{predict_class_train}
\alias{predict_reg_train}
\alias{predict_class_cv}
\alias{predict_reg_cv}
\alias{predict_class_test}
\alias{predict_reg_test}
\title{Predict train data outcome.}
\usage{
predict_class_train(class_model)

predict_reg_train(reg_model)

predict_class_cv(class_model)

predict_reg_cv(reg_model)

predict_class_test(class_model, newdata)

predict_reg_test(reg_model, newdata)
}
\arguments{
\item{class_model}{a \code{caretx} classification model.}

\item{reg_model}{a \code{caretx} regression model.}

\item{newdata}{a data frame with test data.}
}
\value{
a \code{predx} object.
}
\description{
a helper function extracting the predictions for the train, cross-validation
and test data set from a classification or regression \code{caretx} model.
}
