% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prediction.R
\name{predict_reg_cv}
\alias{predict_reg_cv}
\title{Predict cross-validation data outcome for regression models.}
\usage{
predict_reg_cv(reg_model)
}
\value{
a predx object.
}
\description{
a helper function extracting the predictions for the cross-validation data set from
a regression caretx model.
}
