% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prediction.R
\name{predict_reg_test}
\alias{predict_reg_test}
\title{Predict test data outcome for regression models.}
\usage{
predict_reg_test(reg_model, newdata)
}
\value{
a predx object.
}
\description{
a helper function extracting the predictions for the test data set from
a regression caretx model.
}
