% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prediction.R
\name{predict_reg_train}
\alias{predict_reg_train}
\title{Predict train data outcome for regression models.}
\usage{
predict_reg_train(reg_model)
}
\value{
a predx object.
}
\description{
a helper function extracting the predictions for the train data set from
a regression caretx model.
}
