% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_oop.R
\name{print.predx}
\alias{print.predx}
\title{Print a predx object.}
\usage{
\method{print}{predx}(predx_object)
}
\arguments{
\item{predx_object}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}
}
\value{
nothing, called for its side effects.
}
\description{
prints the heading of the prediction table.
}
