% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{residuals.caretx}
\alias{residuals.caretx}
\alias{residuals.predx}
\title{Calculate model or prediction residuals.}
\usage{
\method{residuals}{caretx}(object, newdata = NULL, ...)

\method{residuals}{predx}(object, ...)
}
\arguments{
\item{object}{\code{\link{caretx}} model or
\code{\link{predx}} prediction object.}

\item{newdata}{test data set.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of data frames for the test, resample (CV) and
training data set prediction with residuals and potential outliers.
}
\description{
Computes extended working residuals of prediction and potential outliers.
}
\details{
Calculates working residuals for regression and binary
classification predictions.
In addition, squared and standardized residuals are returned along with
expected normal distribution values for the standardized residuals
and the true outcome.
Potential outliers are identified by the two-SD criterion.
Returns NULL for multi-class models.
}
