% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_oop.R
\name{residuals.predx}
\alias{residuals.predx}
\title{Extract working residuals from a predx object.}
\usage{
\method{residuals}{predx}(object, ...)
}
\arguments{
\item{object}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}
}
\value{
a data frame with residuals.
}
\description{
Extracts extended working residuals of prediction and potential outliers.
}
\details{
Calculates working residuals for regression and binary classification predictions.
In addition, squared and standardized residuals are returned along with
expected normal distribution values for the standardized residuals and the true outcome.
Potential outliers are identified by the two-SD criterion.
}
