% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{split.caretx}
\alias{split.caretx}
\title{Split predictions by an explanatory factor.}
\usage{
\method{split}{caretx}(x, f, drop = FALSE, newdata = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{caretx}} object.}

\item{f}{a splitting factor, in quoted or unquoted form.}

\item{drop}{logical, should unused levels of the splitting factor
f be dropped?}

\item{newdata}{an optional data frame with the test data.}

\item{...}{extra arguments, currently none.}
}
\value{
a plain list of \code{\link{predx}} objects.
}
\description{
The \code{split()} method for the \code{\link{caretx}} class generates predictions
for the training, resample (CV) and, optionally, test data set and splits
them by the levels of an explanatory factor present in the training data.
}
\details{
This method may be used to investigate quality of prediction in a particular
subset or subsets of the data set. The method returns a plain list of
\code{\link{predx}} objects, whose properties can be further explored
with the specific \code{\link{summary.predx}} and \code{\link{plot.predx}}
methods.
}
