% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{squared}
\alias{squared}
\alias{squared.predx}
\alias{squared.caretx}
\title{Square distances from the outcome.}
\usage{
squared(x, ...)

\method{squared}{predx}(x, ...)

\method{squared}{caretx}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{\code{\link{caretx}} model or \code{\link{predx}} prediction object.}

\item{...}{extra arguments passed to methods, currently none.}

\item{newdata}{test data set (optional).}
}
\value{
For \code{predx} objects, a data frame with
the columns \code{.outcome} and \code{.fitted} representing the true outcome
and predicted class assignment, respectively, and \code{square_dist} with the
squared distance from the outcome.
}
\description{
Calculates square distances from the true outcome, i.e. squared differences
between the outcome and fitted class probability as defined in the formula
of the Brier score.
}
\details{
The sum of squared distances between fitted and outcome over
all observations is the Brier score, which ranges from 0 to 1 for binary
classification models and from 0 to 2 for multi-class classifiers.
In each case, 0 indicates a perfect concordance, while 1 or 2
(binary or multi level) hallmarks a completely false prediction.
For regression, squared working residuals are returned.
\code{squared()} is a S3 generic function.
}
\references{
Brier GW. VERIFICATION OF FORECASTS EXPRESSED IN TERMS OF PROBABILITY.
Mon Weather Rev (1950) 78:1–3.
doi:10.1175/1520-0493(1950)078<0001:vofeit>2.0.co;2

Goldstein-Greenwood J. A Brief on Brier Scores | UVA Library. (2021) Available
at: https://library.virginia.edu/data/articles/a-brief-on-brier-scores
}
