% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{summary.caretx}
\alias{summary.caretx}
\title{Get the fit statistic summary of a caretx object.}
\usage{
\method{summary}{caretx}(
  object,
  newdata = NULL,
  ci = FALSE,
  cv_ci_method = c("percentile", "bca", "norm"),
  plain = FALSE
)
}
\arguments{
\item{object}{caretx model.}

\item{newdata}{test data set.}

\item{ci}{logical, should 95\% CI be calculated for non-CV predictions? ignored for CV prediction.}

\item{cv_ci_method}{method used for calculation of 95\% CI for the CV folds: normal distribution, percentile
or BCA (\code{\link[coxed]{bca}}). Defaults to 'percentile'.}

\item{plain}{logical, should the results be coerced to a single data frame?}
}
\value{
a list of data frames with the fit summary statistic or a prediction data frame if plain is set to TRUE.
}
\description{
Calculates fit statistics of a caretx object in the training and cross-validation data set.
If newdata is provided, fit statistics are computed for the test data as well.
For regression those include the fit errors and outcome:fitted correlations. For binary classification
concordance (C-index), kappa, accuracy and receiver-operator characteristic (ROC) is returned. For multi-class
classification, only the accuracy and kappa are currently available.
}
\details{
Harrell's C-index is calculated as specified for \code{\link[survival]{concordance}}. Unweighted Cohen's kappa
is obtained as described for \code{\link[vcd]{Kappa}}. ROC is computed for prediction probabilities
using \code{\link[caret]{twoClassSummary}}. Fit errors are based on working residuals and include mean absolute error (MAE),
mean squared error (MSE) and root-MSE (RMSE). Pseudo R squared is calculated as 1 - MSE/Var(y). Caret R squared is calulated
as root square of the Pearson correaltion coefficient between the fitted and observed values. Pearson correlation is
obtained with \code{\link[base]{cor.test}}, Spearman correlation is computed with \code{\link[DescTools]{SpearmanRho}},
Kendall's TauB is obtained with \code{\link[DescTools]{KendallTauB}}.\cr
For cross-validation (CV) prediction, statistic values are calculated as mean across the CV with 95\% confidence intervals (CI).
For non-CV predictions, 95\% CI are obtained as an option by the specific downstream functions.
}
