% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{analyze_reg}
\alias{analyze_reg}
\alias{analyze_class}
\title{Analyze regression, binary or multi-class fit.}
\usage{
analyze_reg(data, ci = FALSE)

analyze_class(data, classes, c_index = TRUE, ci = FALSE)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{ci}{should 95\% CI be computed?}

\item{classes}{a vector with classes of the outcome.}

\item{c_index}{logical, should concordance index be computed?}
}
\value{
a tibble with the statistic values.
}
\description{
For regression: computes fit errors and fit - outcome correlations.
For binary and multiple-class models, C-index,
receiver-operator characteristic, Cohen's kappa and accuracy is retrieved.
Single class-relevant statistics are averaged over all classes in case of
multi-class predictions.
}
