% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{caretx}
\alias{caretx}
\title{Create a caretx object.}
\usage{
caretx(caret_model)
}
\arguments{
\item{train_model}{a caret model (\code{\link[caret]{train}}).}
}
\value{
a caretx object.
}
\description{
converts a caret model (class 'train', see: \code{\link[caret]{train}}) to a caretx object.
}
\details{
The input caret model has to contain a table with the final model predictions
(\code{\link[caret]{trainControl}}, savePredictions = 'final'),
training data (\code{\link[caret]{trainControl}}, returnData  = TRUE)
and final model resamples raining data (\code{\link[caret]{trainControl}}, returnResamp = 'final').
For classification models, the final model probabilities need to be provided
(\code{\link[caret]{trainControl}}, classProbs = TRUE).
Currently designed to work only with cross-validated models (\code{\link[caret]{trainControl}},
method = 'cv' or 'repeatedcv'), generated with a formula.
}
