% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{caretx}
\alias{caretx}
\alias{as_caretx}
\alias{is_caretx}
\title{Create a caretx object.}
\usage{
caretx(caret_model)

as_caretx(caret_model)

is_caretx(x)
}
\arguments{
\item{caret_model}{a caret model (class \code{\link[caret]{train}}).}

\item{x}{an object.}
}
\value{
a \code{caretx} object; \code{is_caretx()} returns a logical value.
}
\description{
Converts a caret model (class \code{\link[caret]{train}}) to
a \code{caretx} object.
}
\details{
The input caret model has to contain a table with the final model
predictions
(\code{\link[caret]{trainControl}}, savePredictions = 'final'),
training data (\code{\link[caret]{trainControl}}, returnData  = TRUE)
and final model resamples raining data (\code{\link[caret]{trainControl}},
returnResamp = 'final').
For classification models, the final model probabilities need to be provided
(\code{\link[caret]{trainControl}}, classProbs = TRUE).
Currently designed to work only with cross-validated models (\code{\link[caret]{trainControl}},
method = 'cv' or 'repeatedcv'), generated with a formula.
Of note, in case of binary classifiers, the 'positive' outcome is
defined by the second level of the outcome variable.
I.e. if \verb{outcome <- factor(x, c('a', 'b')}, the level \code{b} is considered as
the event of interest.
}
\references{
Kuhn M. Building predictive models in R using the caret
package. J Stat Softw (2008) 28:1–26. doi:10.18637/jss.v028.i05
}
