% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{confusion}
\alias{confusion}
\title{Confusion matrix.}
\usage{
confusion(x, ...)
}
\value{
a table object representing a confusion matrix.
}
\description{
S3 generic function used to create confusion matrices.
}
