% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_resids}
\alias{get_resids}
\alias{get_var}
\alias{get_errors}
\alias{get_correlations}
\title{Calculate residuals, variance, errors and correlation of the outcome and fit.}
\usage{
get_resids(data)

get_var(data)

get_errors(data)

get_correlations(data, ci = TRUE)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{ci}{logical, should 95\% confidence intervals be calculated?}
}
\value{
a numeric vector of residuals or a tibble with statistic values.
}
\description{
Calculates working residuals (\code{get_resids()}),
variance of the outcome (\code{get_var()}),
fir errors based on MSE and outcome variance
(\code{get_errors()}: MAE, MSE, RMSE and pseudo-R-squared),
and correlation between the outcome and fitted values
(\code{get_correlations()}: Pearson, Spearman and Kendal TauB).
}
