% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_resids_binary}
\alias{get_resids_binary}
\title{Binary classifier residuals.}
\usage{
get_resids_binary(data, classes)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{classes}{a vector with the class order.}
}
\value{
a tibble with the statistic values.
}
\description{
Retrieves working binary residuals.
}
