% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{get_roc}
\alias{get_roc}
\title{Receiver-operating characteristic.}
\usage{
get_roc(data, classes)
}
\arguments{
\item{data}{a data frame with .outcome and .fitted columns.}

\item{classes}{a vector with the class order.}
}
\value{
a tibble with the statistic values.
}
\description{
A wrapper around \code{\link[caret]{twoClassSummary}} and
\code{\link[caret]{multiClassSummary}}.
}
\details{
For multi-class predictions, stats referring to class detection
(such as sensitivity or specificity) are averaged over all categories.
}
