% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{nobs.caretx}
\alias{nobs.caretx}
\title{Number of complete observations in a caretx model.}
\usage{
\method{nobs}{caretx}(x)
}
\arguments{
\item{x}{caretx model.}
}
\value{
number of complete cases used for modeling.
}
\description{
The number of complete observations.
}
