% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_oop.R
\name{plot.predx}
\alias{plot.predx}
\title{Plot diagnostic plots and predictions versus outcome for predx objects.}
\usage{
\method{plot}{predx}(
  x,
  type = c("diagnostic", "fit", "regression", "roc", "confusion"),
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  signif_digits = 2,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{x}{an object of class 'predx_object' created e.g. by \code{\link{predict.caretx}}.}

\item{type}{type of the plot. 'diagnostic': diagnostic plots of residuals, 'fit': the fitted versus outcome plot
specific for the prediction character (regression, ROC for binary classification and heat map of the confusion matrix
for multi-class prediction), 'regression': point plot of the fitted vs outcome values, 'roc': ROC plot, 'confusion':
heat map representation of the confusion matrix.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle, statistics specific for the prediction type, if not specified otherwise by the user.}

\item{plot_tag}{plot tag, number of complete observations and events if not specified otherwise by the user.}

\item{signif_digits}{significant digits used for rounding of statistic values presented in the plot.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{extra arguments passed to plotting functions, as specified by the plot type.
For 'diagnostic': \code{\link{get_qc_plots}}, for 'regression': \code{\link{plot_regression}},
for 'roc': \code{\link{plot_roc}}, for 'confusion': \code{\link{plot_confusion}}.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}
}
\value{
a ggplot with the requested content.
}
\description{
Generates diagnostic plots of residuals or plots of fitted versis outcome
as appropriate for the particular prediction type and displays prediction type-specific analysis results.
}
