% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{predict.caretx}
\alias{predict.caretx}
\title{Predict the outcome for a caretx model.}
\usage{
\method{predict}{caretx}(caretx_model, newdata = NULL, plain = FALSE)
}
\arguments{
\item{caretx_model}{caretx model.}

\item{newdata}{test data set.}

\item{plain}{logical, should the results be coerced to a single data frame?}
}
\value{
a list of predx objects containing the predictions, model and data set type information
or a prediction data frame if plain is set to TRUE.
}
\description{
predicts the outcome in train, cross-validation and, optionally, test data set.
}
\details{
extends the output of \code{\link[caret]{predict.train}} method.
}
