% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prediction.R
\name{predict_class_train}
\alias{predict_class_train}
\title{Predict train data outcome for classification models.}
\usage{
predict_class_train(class_model)
}
\value{
a predx object.
}
\description{
a helper function extracting the predictions for the train data set from
a classification caretx model.
}
