% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{predx}
\alias{predx}
\alias{is_predx}
\title{Create a predx object.}
\usage{
predx(
  data,
  classes = NULL,
  type = c("regression", "binary", "multi_class"),
  prediction = c("train", "cv", "test")
)

is_predx(x)
}
\arguments{
\item{data}{a data frame with predictions, the
columns .outcome and .fitted are obligatory.}

\item{classes}{an optional character vector with class names.}

\item{type}{model type (\code{regression}, \code{binary} or \code{multi_class}).}

\item{prediction}{type of prediction (\code{training}, \code{cv} or \code{test}).}

\item{x}{an object.}
}
\value{
a \code{predx} object; \code{is_predx()} returns a logical value.
}
\description{
Creates a \code{predx} object storing extended predictions of
\code{\link{caretx}} models with information
concerning model type, class number and the type of prediction
(training, cross-validation or test).
}
\details{
The \code{type} argument describes the prediction type:
\code{regression}, \code{binary} or \code{multi_class}.
}
\references{
Kuhn M. Building predictive models in R using the caret
package. J Stat Softw (2008) 28:1–26. doi:10.18637/jss.v028.i05
}
