% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{predx}
\alias{predx}
\title{Create a predx object.}
\usage{
predx(
  data,
  classes = NULL,
  type = c("regression", "binary", "multi_class"),
  prediction = c("train", "cv", "test")
)
}
\arguments{
\item{data}{data frame with predictions, the columns .outcome and .fitted are obligatory.}

\item{type}{model type (regression, binary or multi_class).}

\item{prediction}{type of prediction (training, cross-validation or test).}
}
\value{
a predx object.
}
\description{
creates a predx object storing extended predictions of caretx models (\code{\link{caretx}}) with information
concerning model type, class number and the type of prediction (training, cross-validation or test).
}
\details{
The type argument describes the prediction type: regression, binary or multi_class.
}
