% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_oop.R
\name{print.predx}
\alias{print.predx}
\title{Print a predx object.}
\usage{
\method{print}{predx}(x, ...)
}
\arguments{
\item{x}{a \code{predx} object.}

\item{...}{extra arguments, currently none.}
}
\value{
nothing, called for its side effects.
}
\description{
prints the heading of the prediction table.
}
