% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretx_oop.R
\name{residuals.caretx}
\alias{residuals.caretx}
\title{Extract working residuals from a caretx model.}
\usage{
\method{residuals}{caretx}(caretx_model, newdata = NULL)
}
\arguments{
\item{caretx_model}{caretx model.}

\item{newdata}{test data set.}
}
\value{
a list of data frames with residuals and potential outliers.
}
\description{
Extracts extended working residuals of prediction and potential outliers.
}
\details{
Calculates working residuals for regression and binary classification predictions.
In addition, squared and standardized residuals are returned along with
expected normal distribution values for the standardized residuals and the true outcome.
Potential outliers are identified by the two-SD criterion.
}
